abstract linking binaries
